move handleLogin function to LoginPage
